fix unnecessary onnx model file download (#2191)

+++ src/main/java/io/anserini/search/query/QueryEncoder.java
@@ -22,13 +22,23 @@ public abstract class QueryEncoder {
 
   static protected Path getModelPath(String modelName, String modelURL) throws IOException {
     File modelFile = new File(getCacheDir(), modelName);
-    FileUtils.copyURLToFile(new URL(modelURL), modelFile);
+    if (!modelFile.exists()) {
+      System.out.println("Downloading model");
+      FileUtils.copyURLToFile(new URL(modelURL), modelFile);
+    } else{
+      System.out.println("Model already exists");
+    }
     return modelFile.toPath();
   }
 
   static protected Path getVocabPath(String vocabName, String vocabURL) throws IOException {
     File vocabFile = new File(getCacheDir(), vocabName);
-    FileUtils.copyURLToFile(new URL(vocabURL), vocabFile);
+    if (!vocabFile.exists()) {
+      System.out.println("Downloading vocab");
+      FileUtils.copyURLToFile(new URL(vocabURL), vocabFile);
+    } else{
+      System.out.println("Vocab already exists");
+    }
     return vocabFile.toPath();
   }
